DA-643 add dry run logic to import (#711)
* fix typo

* refactor to use = instead of in

* fix typo

* remove logging

* add dry run

* add dry run logging

+++ rest-api/tools/import_organizations.py
@@ -127,7 +127,7 @@ class HPOImporter(CsvImporter):
 
     session.execute(sql)
 
-  def _cleanup_old_entities(self, session, row_list):
+  def _cleanup_old_entities(self, session, row_list, dry_run):
     self.hpo_dao = HPODao()
     existing_hpos = set(hpo.name for hpo in self.hpo_dao.get_all())
     hpo_group_list_from_sheet = [row[HPO_AWARDEE_ID_COLUMN].upper() for row in row_list]
@@ -138,11 +138,12 @@ class HPOImporter(CsvImporter):
       for hpo in hpos_to_remove:
         old_hpo = self.hpo_dao.get_by_name(hpo)
         if old_hpo:
-          logging.info('Deleting old HPO no longer in Google sheet: %s', old_hpo.name)
+          log_prefix = '(dry run) ' if dry_run else ''
+          logging.info(log_prefix + 'Deleting old HPO no longer in Google sheet: %s', old_hpo.name)
           hpo_id_list.append(old_hpo.hpoId)
           self.deletion_count += 1
 
-      if hpo_id_list:
+      if hpo_id_list and not dry_run:
         self.delete_sql_statement(session, hpo_id_list)
 
 
@@ -188,7 +189,7 @@ class OrganizationImporter(CsvImporter):
 
     session.execute(sql)
 
-  def _cleanup_old_entities(self, session, row_list):
+  def _cleanup_old_entities(self, session, row_list, dry_run):
     self.org_dao = OrganizationDao()
     existing_orgs = set(str(org.externalId) for org in self.org_dao.get_all())
     org_group_list_from_sheet = [row[ORGANIZATION_ORGANIZATION_ID_COLUMN].upper()
@@ -198,12 +199,13 @@ class OrganizationImporter(CsvImporter):
     if orgs_to_remove:
       org_id_list = []
       for org in orgs_to_remove:
-        logging.info('Deleting old Organization no longer in Google sheet: %s', org)
+        log_prefix = '(dry run) ' if dry_run else ''
+        logging.info(log_prefix + 'Deleting old Organization no longer in Google sheet: %s', org)
         old_org = self.org_dao.get_by_external_id(org)
         org_id_list.append(old_org.organizationId)
         self.deletion_count += 1
 
-      if org_id_list:
+      if org_id_list and not dry_run:
         self.delete_sql_statement(session, org_id_list)
 
 
@@ -310,7 +312,7 @@ class SiteImporter(CsvImporter):
 
     session.execute(sql)
 
-  def _cleanup_old_entities(self, session, row_list):
+  def _cleanup_old_entities(self, session, row_list, dry_run):
     self.site_dao = SiteDao()
     existing_sites = set(site.googleGroup for site in self.site_dao.get_all())
     site_group_list_from_sheet = [str(row[SITE_SITE_ID_COLUMN].lower()) for row in row_list]
@@ -319,12 +321,13 @@ class SiteImporter(CsvImporter):
     if sites_to_remove:
       site_id_list = []
       for site in sites_to_remove:
-        logging.info('Deleting old Site no longer in Google sheet: %s', site)
+        log_prefix = '(dry run) ' if dry_run else ''
+        logging.info(log_prefix + 'Deleting old Site no longer in Google sheet: %s', site)
         old_site = self.site_dao.get_by_google_group(site)
         site_id_list.append(old_site.siteId)
         self.deletion_count += 1
 
-      if site_id_list:
+      if site_id_list and not dry_run:
         self.delete_sql_statement(session, site_id_list)
 
   def _insert_new_participants(self, entity):
